// src/app/core/models/group.enum.ts
export enum ESurveyTagColor {
  magenta = 'magenta',
  red = 'red'
}

export enum EResponseColor {
  black = 'black',
  gray = 'gray'
}

export enum ETooltipColor {
  gray = 'gray',
  blue = 'blue'
}

export enum EProgressBarColor {
  magenta = 'magenta',
  red = 'red'
}

export enum ESurveyTagText {
  target_met = 'Target Met',
  target_not_met = 'Target Not Met',
  no_responses_collected = 'No Responses Collected'
}